feat: remove deleted bills from bank operations
This PR depends on https://github.com/konnectors/libs/pull/324 which cleans duplicated bills
The goal here is to also clean those bills from linked bank operations
This is at the moment only activated when option "removeDuplicate" is set to true to allow us to
test it heavily before deploying it in production.
The cleaning is done in "linkBankOperations" for performance reasons and avoid getting all the
operation multiple times.
This needs a review from the gangsters especially