fix(konnector-libs): Register the client onne time in local model
Registering the client inside the `localModel` function could cause some
problems if we need to run the model more than one time. Since
`Document.registerClient` can be run only one time, we run it at the
root of the module, so the client is registered for the whole life of
the module.