feat: now use cozy-client's createClientInteractive
This removes cozy-client-js from cozy-jobs-cli as a direct dependency and prepares migration to
cozy-client in cozy-konnector-libs.
Now a cozy-client instance is created and the cozy-client-js instance is created from it in
cozy-konnector-libs.
konnector-dev and run-dev depend on https://github.com/cozy/cozy-client/pull/655 to be published
to work as expected