feat: Allow saveIdentity to handle full identtity and not only contacts
To not change the signature of the function completely, we check if the
first parameter contains a "contact", if not, the object will be wrapped
in a full identity object ({contact: param}.

With this, we suppose that all the identities will contain at least some
contact information which, I suppose is a safe asumption.

I also added some unit test to show the difference.