fix: Use the right mimeTypes function for content type
If the filename would contain / char, the returned content type was the
filename.
A not expected contentType can cause a lot of problems in our
infrastructure.

Instead, we now pass the extension of the file to the contentType
function

In parallel, a fix is done in the stack to validate the content type
properly.