feat: Allow connectors to use this proposed default webpack configuration
This will make easier to push new webpack configuration updates in the
future.

This use this webpack configuration, the connector will need to have

```javascript
module.exports = require('cozy-konnector-build/webpack.config')
```

as the content of its own webpack configuration file. This way, it is
still possible to override the default webpack configuration when needed