fix: No error when assets and travis conf missing
Do not trigger webpack error when the assets directory and travis
configuration file are missing. Some conectors like banking connectors
do not use travis or assets directory. And this is a rule which is
harder to override with webpack-merge.

I prefer to keep the other files mandatories to detect configuration
errors sooner.