feat: Handle sub path in launcher saveFiles
This will allow saveFiles to create a file directly in a sub directory
of the destination folder.

This is espacially used when a konnector needs to fetch multiple
contracts for a given account

This feature was deactivatedin clisk konnector because we did not have a
`mkdirp` implementation yet. Now since cozy-client has already it's own
implementation, we can use it directly.

This feature is needed for the edf clisk konnector to handle multiple
contracts.