feat: Download files progressively only when needed
Now the launcher will only have to call saveFiles directly for all the
given files,  which will avoid to download all the files before calling
saveFiles. (PR in flagship app to follow)

The downloadAndFormatFile option callback has been added to allow
saveFiles to download the corresponding file in the worker when needed

The fetch of the existing files index is now done by the Launcher and
the same index is used for every call to saveFiles of saveBills in the
same konnector run, for performance reason.