feat: Add blockWorkerInteractions command in content script
It is now possible to block user interactions on the worker.
This is needed when the user has validated a login form but when the
konnector does not knoow yet if the login is successful. This avoids
unwanted user interactions which may alter the konnector execution.

Just call :

```javascript
this.blockWorkerInteractions()
this.unblockWorkerInteractions()
```

from the content script