fix: Scope issue
cozy-konnector-dev was not taking the "verbs" attribute to
generate the OAuth client scope. The result was that when
we launched a service by using:
```
yarn run cozy-konnector-dev -m build/manifest.webapp build/services/recurringPurposes/coachco2.js
```

the generated token has all the rights to the permissions and
it was then working.

But when launching the service from the stack with:

```
cozy-stack services run XXXX
```

then it failed because we didn't had the right verbs to the manifest

Here, we fix this issue and cozy-konnector-dev is now reading the
verbs in order to generate the right token.