feat: Add waitForDomReady method to ContentScript
This method will be called when the DOMContentLoaded event is emitted or
if it has already been emitted.

Since this is a pattern we begin to see in every clisk konnectors, let's
avoid to each konnector to implement it :

```javascript
async onWorkerReady() {
  await this.waitForDomReady()
  document.body.addEventListener('click', ...)
}
```

We still add a timeout of 10s to this method, with a warning message (no
error or Exception), to be able to trace any missed event.