fix: ContentScript.waitForDomReady on android
This method did not always work on android and the promise was resolved
only after the timeout of 10s.

This was caused by the document.readyState which could also be
'interactive'. In this case, DOMContentLoad is not supposed to be
already triggered but we cauld miss it while setting the
DOMContentLoaded listener

Since what we are really waiting for in an interactive page, this fix is
correct.